# server/openapi/openapi.baseline.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks/dispatch": {
      "post": {
        "operationId": "TasksController_dispatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispatchTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}/run": {
      "post": {
        "operationId": "TasksController_runNow",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
    "/tasks/{id}/retry": {
      "post": {
        "operationId": "TasksController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/billing/subscriptions/assign": {
      "post": {
        "operationId": "BillingController_assign",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/subscriptions/renew": {
      "post": {
        "operationId": "BillingController_renew",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/subscriptions/{tenantId}": {
      "get": {
        "operationId": "BillingController_subscription",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "GigPayday API",
    "description": "GigPayday platform API contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DispatchTaskDto": {
        "type": "object",
        "properties": {}
      },
...
      "AssignSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
      "RenewSubscriptionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}